Strong_Password.cpp
Before receiving an OTP a user would have usually created a password for the account they are interested to sign up for. With this simple code we can check if the code one can easily check if the password is strong or not. This is a problem I came across a competitive programming platform, so I thought of contributing it to an open source :)